Remove extra style options from the sidebar
